server/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
dateparser
